% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation.r
\name{checkWaterVolumes}
\alias{checkWaterVolumes}
\title{check infiltrated water volumen}
\usage{
checkWaterVolumes(filein, whichCheck, inf_time, cell_volume, infWater_timestep,
  ts_check = 1, plotting = F)
}
\arguments{
\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}

