% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_singleGravityGrid.R
\name{create_singleGravityGrid}
\alias{create_singleGravityGrid}
\title{Create a gravity component grid out of 3 spatially different discretized grids}
\usage{
create_singleGravityGrid(DEM_input_file, DEM_dir, SGloc, grid_discretizations,
  grid_vertical, range_coords_x = NA, range_coords_y = NA,
  grid_radius = NA, correct_SGpillar = NA)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
