% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hygra_internal.r
\name{decimalplaces}
\alias{decimalplaces}
\title{find number of decimal places}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{numerical input.}
}
\description{
find number of decimal places
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@gfz-potsdam.de
}

