% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{demgrid_to_gcompgrid_Edges}
\alias{demgrid_to_gcompgrid_Edges}
\title{Interpolate grid(df) to 3D grid}
\usage{
demgrid_to_gcompgrid_Edges(demgrid_in, grid_discr, depth_split,
  ownSurfaceGrid = F, grid.surface)
}
\arguments{
\item{grid_discr}{vector of discretization of new grid in (x,y,z).}

\item{grid_cords}{coordinates of the original hydrus mesh.}

\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}
}
\description{
interpolate DEM to necessary grid of corresponding layer
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}

