% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{dif_firstLast}
\alias{dif_firstLast}
\title{Calculate differences between last and first value of a vector}
\usage{
dif_firstLast(data_in)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
