% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_viaMinDistance.R
\name{dist}
\alias{dist}
\title{Match data.frames via minimal distance of its (x,y) coordinates}
\usage{
dist(x1, y1, x2, y2)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
demo
track <- data.frame(xt=runif(10,0,360), yt=rnorm(10,-90, 90))
classif <- data.frame(xc=runif(10,0,360), yc=rnorm(10,-90, 90), v1=letters[1:20], v2=1:20)
dist.merge(track, classif, 'xt', 'yt', 'xc', 'yc')
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
