% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{gcomp_testmass}
\alias{gcomp_testmass}
\title{Calculate gravity signal of a test mass at a fixed distance to gravity sensor}
\usage{
gcomp_testmass(mass_loc, mass_geom, gloc, r_inner, r_outer, gama, rho, w)
}
\arguments{
\item{...}{}
}
\description{

}
\examples{
missing 
}
\references{
Marvin Reich (2017), mreich@gfz-potsdam.de
}

