% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{gplot_array}
\alias{gplot_array}
\title{Displaying gravity results}
\usage{
gplot_array(grav.array, grav.name, vertsum = F, layer, cuts = c(-10, -2, -1,
  -0.5, -0.2, -0.1, -0.01, -0.001, 0, 0.001, 0.01, 0.1, 0.2, 0.5, 1, 2, 10))
}
\arguments{
\item{grav.array}{array of values [rows, columns, layers] containing data (usually gravity components or gravity signals), both must have the same dimensions}

\item{grav.name}{name of device used in graphic}

\item{vertsum}{logical, whether values will show vertical sums (TRUE) or only one layer (FALSE), default FALSE}

\item{layer}{layer to show, only affective when vertsum=FALSE}

\item{cuts}{determinating breaks categories of input values}
}
\description{
Visualizes gravity data array (2D)
}
\details{
missing
}
\examples{
example MISSING

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

