% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{gsignal}
\alias{gsignal}
\title{Calulate gravity signals}
\usage{
gsignal(gcompfiles, theta, output = "value")
}
\arguments{
\item{gcompfiles}{vector containing the names (complete paths or relative) of the gravity component files to be used for calculations. These files are generated using gcomp().}

\item{theta}{data.frame with column structure $time, $timestep, $theta (value), $layer}

\item{output}{Defines the output to be a singel value, layer or grid (default is value)
...}
}
\description{
After calculating the gravity components (gcomp()), this function can be used to get real values using soil moisture data (theta). The output of the gravity signal is possible in different dimensions.
}
\details{
Usually one needs relative gravity signals, therefore the inputed theta timeseries should be acutally delta theta vaules per timestep. The
other option is calculating the differences in values per timestep of the output of this function.
}
\examples{
missing 
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

