% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{gsnow}
\alias{gsnow}
\title{Calculate gravity signal from snow storage}
\usage{
gsnow(snow.input)
}
\arguments{
\item{snow.input}{data.frame with columns $time, $SWE (snow water equivalent)}
}
\details{
missing
}
\examples{
missing

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

