% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{gWSC}
\alias{gWSC}
\title{Calculate gravity signal from water storage changes}
\usage{
gWSC(WSC.input, comp.layer)
}
\arguments{
\item{WSC.input}{data.frame with information about $time and further columns with the timeseries of water storage changes in [mm]}

\item{comp.layer}{vector of gravity component of each layer(100\% values), corresponding to a water storage}
}
\description{
Outputs the gravity signal of each storage deparment where water mass changes occured
}
\details{
The number of components has to match the number of columns in WSC.input (+ column $time)
}
\examples{
missing

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

