% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hydrus_gw}
\alias{hydrus_gw}
\title{Data preparation for HYDRUS}
\usage{
hydrus_gw(date.in, date.out, name, depthLB, freq = "hour", aszoo = F)
}
\arguments{
\item{date.in}{starting date (POSIXct)}

\item{date.out}{ending date (POSIXct)}

\item{name}{name of groundwater well to use. Possible so far "BK1", "BK2", "BK3", "BK14".}

\item{depthLB}{depths of lower model boundary. All groundwater level values will be subtracted from this value.}

\item{freq}{frequency of data. Possible values are "day", "hour", "min", "sec".}

\item{aszoo}{output either in data.frame (aszoo=F) or zoo (aszoo=T). Default value is FALSE.}
}
\description{
groundwater TS
}
\details{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

