% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation.r
\name{inf_model}
\alias{inf_model}
\title{Hydro-gravimetric model}
\usage{
inf_model(param_vec)
}
\arguments{
\item{test}{...}
}
\description{
3D distribution algorithm, modeling the soil infiltration of water from a sprinkling experiment
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}

