% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpTS.R
\name{interpTS}
\alias{interpTS}
\title{Interpolate missing time steps of data time series}
\usage{
interpTS(data_in, freq_in, freq_out = NA, aggFunc = "mean", data_col_name)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
