% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.r
\name{kvadose}
\alias{kvadose}
\title{Calculate K unsaturated}
\usage{
kvadose(theta, vgparam, ksat)
}
\arguments{
\item{theta}{soil moisture value which to calculate K unsaturated}

\item{Ksat}{saturated hydraulic conductivity}

\item{thr}{residual moisture content}

\item{ths}{saturated moisture content}

\item{alpha}{inverse of air entry point}

\item{n}{pore size distribution}
}
\description{
Get unsaturated hydraulic conductivities for any soil moisture value based on the van Genuchten parameters.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

