% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation.r
\name{nodestogrid_3d_PP}
\alias{nodestogrid_3d_PP}
\title{Interpolate 3D nodes of a time series to regular grid using parralel computing}
\usage{
nodestogrid_3d_PP(data_input, grid_discr, depth_split, grid_cords = NA,
  nintpmax = 10)
}
\arguments{
\item{data_input}{data to be interpolated. this has to be a data.frame with columns (x, y, z, value).}

\item{grid_discr}{vector of discretization of new grid in (x,y,z).}

\item{grid_cords}{grid to be interpolated to. if no grid is passed, it is created from input_data spatial ranges.}

\item{grid_depth}{vector of vertical extents of new grid in c(min,max).}

\item{nintmax}{maximal number of points to consider in the surroundings for one interpolation point}
}
\description{
interpolate a time series, which has 3 dimensions, using IDW and time-stacking
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@gfz-potsdam.de
}

