% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(x)
}
\arguments{
\item{x}{input dataset}
}
\description{
Normalize data in subtracting the mean and dividing by the standard deviation.
}
\examples{
example MISSING
}
\references{
Marvin Reich (2015), mreich@gfz-potsdam.de
}

