% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{normalize_mean}
\alias{normalize_mean}
\title{Normalize data only sbutracting mean}
\usage{
normalize_mean(x)
}
\arguments{
\item{x}{input dataset}
}
\description{
Normalize data in subtracting the mean and dividing by the standard deviation.
}
\examples{
example MISSING
}
\references{
Marvin Reich (2015), mreich@gfz-potsdam.de
}

