% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hygra_internal.r
\name{PFtoSuc}
\alias{PFtoSuc}
\title{Convert pf-values to suction potential (cm water column)}
\usage{
PFtoSuc(PF)
}
\arguments{
\item{PF}{input PF-values. List, vector.
...}
}
\description{

}
\details{
missing
}
\examples{
missing

}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}

