% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_soildata}
\alias{plot_soildata}
\title{Plot soil data}
\usage{
plot_soildata(file, show, top, bottom, plottogether = F)
}
\arguments{
\item{file}{dataset of measured values as a zoo object}

\item{show}{profile (data column) to plot}

\item{top}{limiting data values at top}

\item{bottom}{limiting data values at bottom}

\item{plottogether}{plot all data in one single plot (TRUE) or subplots for each timeseries (FALSE; default)}
}
\description{
Generally designed for plotting soil moisture data of different sensor profiles
}
\details{
missing
}
\examples{
missing

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

