% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation.r
\name{plotInfWater}
\alias{plotInfWater}
\title{plot irrigated water volume per layer and timestep}
\usage{
plotInfWater(filein, plotvar, showlegend = T)
}
\arguments{
\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}

