% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation.r
\name{plotIrrigationProfile}
\alias{plotIrrigationProfile}
\title{plot irrigated cut profile at certain time step}
\usage{
plotIrrigationProfile(filein, tstep = 360, cutrel = 2, plotvar = "x",
  showlegend = T)
}
\arguments{
\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}

