% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hygra_internal.r
\name{split_ts}
\alias{split_ts}
\title{Split timeseries}
\usage{
split_ts(data.in, n)
}
\arguments{
\item{data.in}{Dataset to be split (must be zoo-object)}

\item{n}{indicates minimum length of splits
...}
}
\description{
Splits a timeseries into differently long pieces of the data, according to the split-indices used.
By default it is intended to split at NA-values. Afterwards splits can be sorted out which don't meet a specified length.
}
\details{
missing
}
\examples{
missing

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

