% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{subtract_maxrad_gcomp}
\alias{subtract_maxrad_gcomp}
\title{subtract everything out of certain radius from gcomp (raw) grid}
\usage{
subtract_maxrad_gcomp(grid_input, gloc, max_rad)
}
\arguments{
\item{grid_input}{grid from where to limit gcomp to a certain radius. Has to be data.frame with columns (x,y,z).}

\item{gloc}{location of gravity sensor (x,y,z).}

\item{max_rad}{radius of exclusion in [meters].}
}
\description{

}
\details{
missing
}
\examples{
example MISSING
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

