% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.r
\name{subtract_umbrellaEffect}
\alias{subtract_umbrellaEffect}
\title{subtract umbrella EFFECT from gcomp (raw) grid}
\usage{
subtract_umbrellaEffect(grid_input, u_polygon, u_depth, sgz)
}
\arguments{
\item{g_grid}{grid for which to calculate gravity components. (generally derived from DEM of region in combination with hydrological modeling mesh/grid).}

\item{senloc}{coordinates of gravity sensor location. data.frame with x,y,z columns.}

\item{g_discr}{discretization of g_grid in differences in x,y,and z-direction. they have to be uniform steps!}

\item{edge}{if grid to calulate is located on a domain edge (top or bottom). for top set edge='first', for bottom set edge='last'.}
}
\description{
Calculates the gravity components (100percent values) for a given DEM using a nested approach
}
\details{
missing
}
\examples{
example MISSING
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

