% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.r
\name{suction}
\alias{suction}
\title{Calculate suction potential from soil moisture content}
\usage{
suction(theta, vgparam)
}
\arguments{
\item{theta}{Soil moisture value (decimal number) to be converted to suction pressure}

\item{vgparam}{data.frame of van Genuchten parameters with the column structure: $thr, $ths, $alpha, $n}
}
\description{
test
}
\details{
missing
}
\examples{
missing

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

