% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.r
\name{template}
\alias{template}
\title{test}
\usage{
template()
}
\arguments{
\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

